backend/requirements.txt
boto3>=1.34.0
requests>=2.31.0
google-genai>=0.3.0